pgxs: Encourage use of 'vpath' directives instead of VPATH
Using a full VPATH can lead to surprising and confusing results, but
it's still desirable to allow extensions to build from outside their
source trees. Make minor pgxs.mk changes to make this easier and
document how to do it.

Most of PostgreSQL has already eliminated VPATH use, but there was
a lingering use in the PGXS logic to copy regression test .data files
to the build tree. Change it to test whether the srcdir and build dir
are the same, so it fires even if the actual VPATH variable is not set.

Add a vpath directive to allow the existing implicit %c: %o rule
used by OBJS to search $(srcdir)/ for .c input files. This is enough
to make most extensions build fine using

   mkdir ../extbuild
   cd ../extbuild/
   make -f ../extsrc/Makefile

without any explicit VPATH= setting.